-- added more asserts -- switched to ZeroMemory(...) instead of = {}; -- released the buffers of the views correctly -- initialized a rasterizer state -- created a viewport and bound it to the rasterizer state -- fixed a bug in the obj to mesh converter -- eliminated some compiler warnings by fixing possible truncation errors -- introduced new global pointers to rendering constructs -> might want to put this into a struct or something -- integrated the obj model loading of a test .obj-file (360.000+ vertices) and its conversion into a mesh object

+++ renderhub_types.h
@@ -125,7 +125,7 @@ struct Mesh
 		TRIANGLE_LIST = 4,
 		TRIANGLE_STRIP = 8,
 	};
-	uint32 vertex_count;
+	uint64 vertex_count;
 	vertex* vertices;
 	uint32 flags;
 };
